review_code/modelarts_deploy/model/requirements.txt
numpy
pillow
torch>=1.1
mmcv
pycocotools
torchvision
albumentations>=0.3.2
imagecorruptions


